fix route model binding with custom middleware

+++ src/Services/FlowState/FlowStateManager.php
@@ -11,6 +11,8 @@ use JobMetric\Flow\Events\FlowState\FlowStateUpdateEvent;
 use JobMetric\Flow\Exceptions\FlowInactiveException;
 use JobMetric\Flow\Exceptions\FlowStateInvalidTypeException;
 use JobMetric\Flow\Exceptions\FlowStateStartTypeIsExistException;
+use JobMetric\Flow\Exceptions\FlowStateStartTypeIsNotChangeException;
+use JobMetric\Flow\Exceptions\FlowStateStartTypeIsNotDeleteException;
 use JobMetric\Flow\Exceptions\FlowStatusInvalidException;
 use JobMetric\Flow\Facades\Flow;
 use JobMetric\Flow\Models\FlowState;
@@ -105,6 +107,7 @@ class FlowStateManager
      * @return FlowState
      * @throws FlowStateInvalidTypeException
      * @throws FlowStatusInvalidException
+     * @throws FlowStateStartTypeIsNotChangeException
      */
     public function update(int $flow_state_id, array $data = []): FlowState
     {
@@ -115,7 +118,11 @@ class FlowStateManager
                 throw new FlowStateInvalidTypeException($data['type']);
             }
 
-            $flowState->type = $data['type'];
+            if($flowState->type == TableFlowStateFieldTypeEnum::START() && $data['type'] != TableFlowStateFieldTypeEnum::START()) {
+                throw new FlowStateStartTypeIsNotChangeException($flowState->flow->driver);
+            } else {
+                $flowState->type = $data['type'];
+            }
         }
 
         if (isset($data['color'])) {
@@ -151,12 +158,17 @@ class FlowStateManager
      * @param int $flow_state_id
      *
      * @return FlowState
+     * @throws FlowStateStartTypeIsNotDeleteException
      */
     public function delete(int $flow_state_id): FlowState
     {
         // todo: check dependencies
 
         $flow_state = $this->show($flow_state_id);
+
+        if($flow_state->type == TableFlowStateFieldTypeEnum::START) {
+            throw new FlowStateStartTypeIsNotDeleteException;
+        }
 
         $flow_state->delete();
